# Conflicts:
#	.travis.yml

+++ tests/Distortion/SD1OverdriveFilter.cpp
@@ -30,20 +30,20 @@ constexpr gsl::index PROCESSSIZE = 1000;
 
 BOOST_AUTO_TEST_CASE( SD1OverdriveFilter_sinus_drive_test )
 {
-  ATK::SD1OverdriveFilter<float> filter;
+  ATK::SD1OverdriveFilter<double> filter;
   filter.set_drive(0.5);
   BOOST_CHECK_EQUAL(filter.get_drive(), 0.5);
 }
 
 BOOST_AUTO_TEST_CASE( SD1OverdriveFilter_sinus_drive_range_test )
 {
-  ATK::SD1OverdriveFilter<float> filter;
+  ATK::SD1OverdriveFilter<double> filter;
   BOOST_CHECK_THROW(filter.set_drive(1.0000001), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( SD1OverdriveFilter_sinus_drive_range2_test )
 {
-  ATK::SD1OverdriveFilter<float> filter;
+  ATK::SD1OverdriveFilter<double> filter;
   BOOST_CHECK_THROW(filter.set_drive(-0.0000001), std::out_of_range);
 }
