+++ tests/Dynamic/GainColoredExpanderFilter.cpp
@@ -17,56 +17,56 @@ constexpr gsl::index PROCESSSIZE = 64;
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_softness_test )
 {
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter;
   filter.set_softness(0.5);
   BOOST_CHECK_EQUAL(filter.get_softness(), 0.5);
 }
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_softness_range_test )
 {
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter;
   BOOST_CHECK_THROW(filter.set_softness(-0.000001), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_color_test )
 {
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter;
   filter.set_color(0.5);
   BOOST_CHECK_EQUAL(filter.get_color(), 0.5);
 }
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_quality_test )
 {
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter;
   filter.set_quality(0.5);
   BOOST_CHECK_EQUAL(filter.get_quality(), 0.5);
 }
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_quality_range_test )
 {
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter;
   BOOST_CHECK_THROW(filter.set_quality(0), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_1_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
   
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(10);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -80,22 +80,22 @@ BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_1_test )
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_0_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 0;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -103,31 +103,31 @@ BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_0_test )
 
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
-    BOOST_REQUIRE_SMALL(outdata[i], 0.001f);
+    BOOST_REQUIRE_SMALL(outdata[i], 0.001);
   }
 }
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_1_threshold_2_ratio_2_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(2);
   filter.set_ratio(2);
   filter.set_softness(1);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -141,25 +141,25 @@ BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_1_threshold_2_ratio_2_test
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_1_threshold_2_ratio_4_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(2);
   filter.set_ratio(4);
   filter.set_softness(1);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -173,25 +173,25 @@ BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_const_1_threshold_2_ratio_4_test
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_always_more_1_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = i/1024.;
   }
   
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
   
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
   
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(1);
   filter.set_quality(.1);
   filter.set_color(.1);
   
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
   
@@ -205,25 +205,25 @@ BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_always_more_1_test )
 
 BOOST_AUTO_TEST_CASE( GainColoredExpanderFilter_always_less_1_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = i/1024.;
   }
   
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
   
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
   
-  ATK::GainFilter<ATK::GainColoredExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainColoredExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(1);
   filter.set_quality(.1);
   filter.set_color(-.1);
   
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
